return controller if body lacks values

+++ backend/src/goals/goals.controller.ts
@@ -10,18 +10,23 @@ export class GoalsController {
 
   @Get("getUserGoals")
   async getUserGoals(@Query() query) {
+    if(!query.session) return;
     const result = await this.goalsService.getUserGoals(query.session);
     return result;
   }
 
   @Get("getUserGoal")
   async getUserGoal(@Query() query) {
+    if(!query.session) return;
+    if(!query.goal_id) return;
     const result = await this.goalsService.getUserGoal(query.session, query.goal_id);
     return result;
   }
 
   @Post("addUserGoal")
   async addUserGoal(@Body() body){
+    if(!body.session) return;
+    if(!body.goal) return;
     const isUserPremium = await this.stripeService.isUserPremium(body.session);
     const result = await this.goalsService.addUserGoal(body.session, body.goal, isUserPremium)
     return result;
@@ -29,24 +34,31 @@ export class GoalsController {
 
   @Post("updateUserGoal")
   async updateUserGoal(@Body() body){
+    if(!body.session) return;
+    if(!body.goal) return;
     const result = await this.goalsService.updateUserGoal(body.session, body.goal)
     return result
   }
 
   @Delete("deleteUserGoal")
   async deleteUserGoal(@Body() body){
+    if(!body.session) return;
+    if(!body.goal) return;
     const result = await this.goalsService.deleteUserGoal(body.session, body.goal_id)
     return result;
   }
 
   @Get("getUserUpdates")
   async getUserUpdates(@Query() query) {
+    if(!query.session) return;
     const result = await this.goalsService.getUserUpdates(query.session);
     return result;
   }
 
   @Get("getGoalUpdates")
   async getGoalUpdates(@Query() query) {
+    if(!query.session) return;
+    if(!query.goal_id) return;
     const result = await this.goalsService.getGoalUpdates(query.session, query.goal_id);
     return result;
   }
@@ -54,18 +66,24 @@ export class GoalsController {
 
   @Post("addUserUpdate")
   async addUserUpdate(@Body() body){
+    if(!body.session) return;
+    if(!body.update) return;
     const result = await this.goalsService.addUserUpdate(body.session, body.update)
     return result;
   }
 
   @Post("updateUserUpdate")
   async updateUserUpdate(@Body() body){
+    if(!body.session) return;
+    if(!body.update) return;
     const result = await this.goalsService.updateUserUpdate(body.session, body.update)
     return result
   }
 
   @Delete("deleteUserUpdate")
   async deleteUserUpdate(@Body() body){
+    if(!body.session) return;
+    if(!body.update_id) return;
     const result = await this.goalsService.deleteUserUpdate(body.session, body.update_id)
     return result;
   }
